+++ test/libyul/YulOptimizerTestCommon.cpp
@@ -407,7 +407,12 @@ YulOptimizerTestCommon::YulOptimizerTestCommon(std::shared_ptr<Object const> _ob
 			{
 				Object object(*m_optimizedObject);
 				object.setCode(std::make_shared<AST>(*m_object->dialect(), std::get<Block>(ASTCopier{}(block))));
-				block = std::get<1>(StackCompressor::run(object, true, maxIterations));
+				block = std::get<1>(StackCompressor::run(
+					object,
+					true,
+					maxIterations,
+					frontend::OptimiserSettings::standard().expectedExecutionsPerDeployment
+				));
 			}
 			BlockFlattener::run(*m_context, block);
 			return block;